requirements.txt
azure-storage-blob
mysql-connector-python
numpy
pandas
torch
tqdm
